pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nessus_verifier"
version = "0.0.1"
description = "Pentests services from nessus file"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "kcancurly" }
]
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
    "argcomplete",
    "requests",
    "bs4",
    "dnspython",
    "impacket",
    "pyaml",
    "pysmb",
    "pymongo",
    "python-nmap",
    "rich",
    "psycopg",
    "paramiko",
    "toml",
    "stomp.py",
    "pymssql",
    "asyncssh[bcrypt]",
    "colorama",
    "pymysql",
    "packaging",
    "pika",
    "redis[hiredis]",
    "BAC0",
    "nvdlib",
    "cidr_man",
    "selenium",
    "html5lib",
    "python-i18n",
    "paho-mqtt",
    "xmpppy",
    "jmxquery",
    "openpyxl",
    "oracledb",
    "pymodbus",
    "smpplib @ git+https://github.com/python-smpplib/python-smpplib.git",
    "tftpy",
    "loguru",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/kcancurly/nessus-verifier"

[project.scripts]
nv-parse = "src.modules.nv_parse:main"
nv-solver = "src.solvers.solver:main"
nv-nmap = "src.modules.nv_nmap:main"
nv-nuclei = "src.modules.nv_nuclei:main"
nv-ncat = "src.modules.nv_ncat:main"
nv-service = "src.services.nv_services:main"
nv-url = "src.url.url:main"
nv-nessus = "src.modules.nv_nessus:main"
nv-misc = "src.modules.nv_misc:main"
nv-evidence = "src.modules.nv_evidence:main"

[tool.setuptools]
packages = [
    "src",
    "src.modules",
    "src.modules.parse_post",
    "src.solvers",
    "src.services",
    "src.url",
    "src.utilities",
]

